[Refactor] CardData.swift의 열거형과 Card.swift의 구조체를 하나로 통합 Card 구조체 안에 CardData 열거형을 선언 (CardData 열거형은 Card 타입 이외에서는 쓰이지 않으므로)
