+++ PokerGameApp/PokerGameApp/Card.swift
@@ -9,25 +9,68 @@ import Foundation
 
 struct Card {
     
-    //다른 객체에 의해 카드값이 변경되면 안되므로 let으로 선언했습니다.
-    //Card 구조체에서 제공해주는 방식으로만 faceData와 numericData에 접근할 수 있도록 private으로 선언했습니다.
-    //다른 객체는 Card 객체의 faceData 값에만, 혹은 numericData 값에만 개별적으로 접근할 수 없고,
-    //아래에 선언된 Card 구조체가 제공해주는 extension의 프로퍼티를 통해서
-    //faceData와 numericData가 결합되어 표현된 하나의 String 형식으로만 받아볼 수 있게하기 위함입니다.
+    enum CardData {
+        enum FaceData: String, CustomStringConvertible {
+            case spade
+            case club
+            case heart
+            case diamond
+            
+            var description: String {
+                switch self {
+                case .spade:
+                    return "♠"
+                case .club:
+                    return "♣"
+                case .heart:
+                    return "♥"
+                case .diamond:
+                    return "♦"
+                }
+            }
+        }
+
+        enum NumericData: Int, CustomStringConvertible {
+            case ace = 1
+            case two
+            case three
+            case four
+            case five
+            case six
+            case seven
+            case eight
+            case nine
+            case ten
+            case jack
+            case queen
+            case king
+            
+            var description: String {
+                switch self {
+                case .ace:
+                    return "A"
+                case .two, .three, .four, .five, .six, .seven, .eight, .nine, .ten:
+                    return "\(self.rawValue)"
+                case .jack:
+                    return "J"
+                case .queen:
+                    return "Q"
+                case .king:
+                    return "K"
+                }
+            }
+        }
+    }
+    
     private let faceData: CardData.FaceData
     private let numericData: CardData.NumericData
     
-    
-    init(faceData: CardData.FaceData, numericData: CardData.NumericData) {
+    init(faceData: Card.CardData.FaceData, numericData: Card.CardData.NumericData) {
         self.faceData = faceData
         self.numericData = numericData
     }
 }
 
-//메소드가 아닌 extension을 통해 구현했습니다.
-//연산 작업이나 특정한 기능을 수행하는 것이 아니고
-//그저 단순히 Card 객체가 갖는 두 프로퍼티의 값을 붙인 String 타입으로 반환해주기만 하면 되므로
-//extension을 통해 프로퍼티 형식으로 바로 접근하는 것이 더 직관적일 것 같다고 생각했습니다.
 extension Card: CustomStringConvertible {
     var description: String {
         return "\(self.faceData)\(self.numericData)"
